package-lock.json Installation of prisma and prisma client.
package.json
Installation of prisma and prisma client.

Untracked Files
Creation and edit of .env file for database URL.
Creation of test index.js file to test querying to database. Was unable to get this to work.
Creation of prisma schema file with test models from prisma setup as well as the model for the existing table on heroku.